# rest/server/create_taco.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	tacomodule "taqueria/core/module/taco"
	"taqueria/core/module/taco/types"
	main_entity "taqueria/entity/taco"
)

func (h *handlers) CreateTaco(w http.ResponseWriter, r *http.Request) {
	var item main_entity.Taco
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.Taco().Insert(r.Context(), types.UpsertRequest{
		Taco: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.Taco().FetchTacoByID(r.Context(), types.FetchTacoByIDRequest(res), tacomodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.ID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}