# rest/server/create_taqueria.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	taqueriamodule "taqueria/core/module/taqueria"
	"taqueria/core/module/taqueria/types"
	main_entity "taqueria/entity/taqueria"
)

func (h *handlers) CreateTaqueria(w http.ResponseWriter, r *http.Request) {
	var item main_entity.Taqueria
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.Taqueria().Insert(r.Context(), types.UpsertRequest{
		Taqueria: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.Taqueria().FetchTaqueriaByID(r.Context(), types.FetchTaqueriaByIDRequest(res), taqueriamodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.ID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}