# rest/server/create_visita.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	visitamodule "taqueria/core/module/visita"
	"taqueria/core/module/visita/types"
	main_entity "taqueria/entity/visita"
)

func (h *handlers) CreateVisita(w http.ResponseWriter, r *http.Request) {
	var item main_entity.Visita
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.Visita().Insert(r.Context(), types.UpsertRequest{
		Visita: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.Visita().FetchVisitaByID(r.Context(), types.FetchVisitaByIDRequest(res), visitamodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.ID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}